Add check if product already exists

+++ addProduct.php
@@ -27,28 +27,35 @@
     <a href="product.php">Alle Produkte anzeigen</a>
 </p>
 <?php
+    $alreadyExist = isset($_POST["product_name"]) && alreadyExistsProduct($_POST["product_name"]);
     if(isset($_GET['add'])) {
-        $conn = new Mysql();
-        $conn -> dbConnect();
-        
-        $NULL = [
-            "type" => "null",
-            "val" => "null"
-        ];
-        
-        $product_name = [
-            "type" => "char",
-            "val" => $_POST["product_name"]
-        ];
-                
-        $data = array($NULL, $product_name);
-        
-        $conn -> insertInto('T_Product', $data);
-        $conn -> dbDisconnect();
-        
-        echo '<div class="infobox">';
-        echo 'Das Produkt <strong>' . $_POST["product_name"] . '</strong> wurde hinzugefügt';
-        echo '</div>';
+        if($alreadyExist) {
+            echo '<div class="warning">';
+            echo 'Das Produkt <strong>' . $_POST["product_name"] . '</strong> existiert bereits';
+            echo '</div>';
+        } else {
+            $conn = new Mysql();
+            $conn -> dbConnect();
+
+            $NULL = [
+                "type" => "null",
+                "val" => "null"
+            ];
+
+            $product_name = [
+                "type" => "char",
+                "val" => $_POST["product_name"]
+            ];
+
+            $data = array($NULL, $product_name);
+
+            $conn -> insertInto('T_Product', $data);
+            $conn -> dbDisconnect();
+
+            echo '<div class="infobox">';
+            echo 'Das Produkt <strong>' . $_POST["product_name"] . '</strong> wurde hinzugefügt';
+            echo '</div>';
+        }
     }
 ?>
 <form action="?add=1" method="POST">
